Group connection helper methods in module + add removeKnownConnections()
