+++ src/components/edges.tsx
@@ -19,7 +19,7 @@ import {
 import { textColor } from "./drivenColors";
 import { cn } from "@/lib/utils";
 import { store } from "@/stores/global";
-import { connection2id } from "@/lib/connection2id";
+import { connection2id } from "@/lib/connectionHelpers";
 import { useCallback } from "react";
 import { conflict2id } from "@/stores/component-network";
 import { XIcon } from "lucide-react";
